pyproject.toml
[project]
name = "ragcli"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "Anchit Bajaj", email = "ab@abifog.com" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "openai>=2.2.0",
    "pydantic-ai-slim[openai]>=1.0.17",
    "qdrant-client>=1.15.1",
    "typer>=0.19.2",
]

[project.scripts]
ragexperiments = "ragcli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
indent-width = 4

[tool.pyright]
typeCheckingMode = "standard"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
]
